# Groupe 1 /DADdy/apps/breezy-api/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Breezy API
  version: 1.0.0
  description: >
    Réseau social léger type Twitter/X. Architecture microservices (Go/Gin)
    derrière une API Gateway, données PostgreSQL, authentification JWT.
    Tous les appels passent par la gateway.
servers:
  - url: http://localhost:8080
    description: API Gateway (développement)
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    patch:
      tags: [Users]
      summary: Éditer son profil (Fx10)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/avatar:
    put:
      tags: [Users]
      summary: Téléverser son avatar (Fx10)
      description: >
        Envoie un fichier image (PNG, JPEG, WebP ou GIF, 5 Mio max) stocké sur MinIO.
        Remplace l'avatar existant le cas échéant.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar mis à jour
          content:
            application/json:
              schema:
                type: object
                required: [avatarUrl]
                properties:
                  avatarUrl:
                    type: string
                    description: URL publique de l'avatar servi par MinIO
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Users]
      summary: Supprimer son avatar (Fx10)
      responses:
        '204':
          description: Avatar supprimé
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/banner:
    put:
      tags: [Users]
      summary: Téléverser sa bannière (Fx10)
      description: >
        Envoie un fichier image (PNG, JPEG, WebP ou GIF, 10 Mio max) stocké sur MinIO.
        Remplace la bannière existante le cas échéant.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Bannière mise à jour
          content:
            application/json:
              schema:
                type: object
                required: [bannerUrl]
                properties:
                  bannerUrl:
                    type: string
                    description: URL publique de la bannière servie par MinIO
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Users]
      summary: Supprimer sa bannière (Fx10)
      responses:
        '204':
          description: Bannière supprimée
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/suggestions:
    get:
      tags: [Users]
      summary: Suggestions d'abonnement (« qui suivre »)
      description: >-
        Comptes à suivre, classés par recouvrement entre les centres d'intérêt du
        lecteur et les thèmes des posts des auteurs (signal interne, jamais exposé
        dans la réponse). Exclut soi, les comptes déjà suivis et les blocages. En
        l'absence de signal thématique (nouvel arrivant), repli sur des comptes
        récents.
      responses:
        '200':
          description: Liste de profils suggérés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    get:
      tags: [Users]
      summary: Profil public d'un utilisateur (Fx10)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Profil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/follow-requests:
    get:
      tags: [Users]
      summary: Demandes d'abonnement en attente reçues (comptes privés)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée des profils demandeurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
  /users/me/follow-requests/{id}/accept:
    post:
      tags: [Users]
      summary: Accepter une demande d'abonnement (id = demandeur)
      description: Crée l'abonnement et notifie le demandeur (follow_request_accepted).
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Demande acceptée, abonnement créé
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/follow-requests/{id}:
    delete:
      tags: [Users]
      summary: Refuser une demande d'abonnement (id = demandeur)
      description: Supprime la demande. Idempotent, aucune notification envoyée.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Demande refusée (ou inexistante)
  /users/{id}/follow:
    post:
      tags: [Users]
      summary: Suivre un utilisateur (Fx9)
      description: >
        Sur un compte public, crée directement l'abonnement (204). Sur un compte
        privé, crée une demande en attente d'acceptation (202).
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Abonnement créé (compte public, idempotent)
        '202':
          description: Demande d'abonnement créée (compte privé)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [requested]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Ne plus suivre (Fx9)
      description: Supprime l'abonnement et toute demande en attente (annulation).
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Abonnement supprimé
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/following:
    get:
      tags: [Users]
      summary: Liste des abonnements (Fx9) — consommé par le Feed service
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée d'utilisateurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
  /users/{id}/followers:
    get:
      tags: [Users]
      summary: Liste des abonnés
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée d'utilisateurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
  /users/{id}/posts:
    get:
      tags: [Posts]
      summary: Posts d'un utilisateur (Fx11 / Fx4)
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée de posts, anti-chronologique
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPosts'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: Identifiant opaque de l'utilisateur, préfixé (ex. "usr_ab12cd34ef56ab78"), émis par l'auth-service.
      schema:
        type: string
        example: usr_ab12cd34ef56ab78
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: content exceeds 280 characters
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHENTICATED
            message: invalid or missing token
    Forbidden:
      description: Authentifié mais non autorisé
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: FORBIDDEN
            message: you are not the author of this post
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: post not found
...
  schemas:
    PostId:
      type: string
      description: Identifiant opaque du post, préfixé par "pst_".
      pattern: '^pst_[a-f0-9]{16}$'
      example: pst_ab12cd34ef56ab78
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
        message:
          type: string
    Pagination:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 137
    UserSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          description: Identifiant opaque préfixé (ex. "usr_ab12cd34ef56ab78"), émis par l'auth-service.
          example: usr_ab12cd34ef56ab78
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
    User:
      type: object
      required: [id, username, role, followersCount, followingCount, createdAt]
      properties:
        id:
          type: string
          description: Identifiant opaque préfixé (ex. "usr_ab12cd34ef56ab78"), émis par l'auth-service.
          example: usr_ab12cd34ef56ab78
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          nullable: true
          description: Renvoyé uniquement lorsque l'utilisateur consulte son propre profil
        displayName:
          type: string
          maxLength: 50
          nullable: true
          description: Nom d'affichage libre, distinct du username
        pronouns:
          type: string
          maxLength: 50
          nullable: true
          description: Pronoms en texte libre (ex. "il/lui", "elle", "they/them")
        bio:
          type: string
          maxLength: 160
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
          description: Image de bannière (en-tête du profil) servie par MinIO
        role:
          type: string
          enum: [user, moderator, admin]
        isPrivate:
          type: boolean
          description: Compte privé — posts/re-posts réservés aux abonnés
        followersCount:
          type: integer
        followingCount:
          type: integer
        isFollowedByMe:
          type: boolean
          description: Contextuel à l'appelant (présent sur GET /users/{id})
        followRequested:
          type: boolean
          description: >
            Contextuel (GET /users/{id}) — true si l'appelant a une demande
            d'abonnement en attente vers ce compte privé
        canViewPosts:
          type: boolean
          description: >
            Contextuel (GET /users/{id}) — false sur un compte privé dont
            l'appelant n'est pas abonné (posts/re-posts masqués)
        createdAt:
          type: string
          format: date-time
    Post:
      type: object
      required: [id, author, content, likesCount, commentsCount, likedByMe, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/PostId'
        author:
          $ref: '#/components/schemas/UserSummary'
        content:
          type: string
          minLength: 1
          maxLength: 280
        tags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        likedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 50
        pronouns:
          type: string
          maxLength: 50
        bio:
          type: string
          maxLength: 160
        isPrivate:
          type: boolean
          description: Bascule le compte en privé/public
...
    PaginatedPosts:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    PaginatedUsers:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'